cloudapi: split into handler and server
These "classes" are huge, let's split them into two files to make the file
smaller.
